fix: way to handle authenticate status
