+++ webapp/frontend/src/containers/NonAuthContainer.tsx
@@ -1,12 +1,22 @@
 import { AppState } from '../index';
-import { Dispatch } from 'redux';
+import { AnyAction } from 'redux';
 import { connect } from 'react-redux';
 import { NonAuthRoute } from '../components/Route/NonAuthRoute';
+import { fetchSettings } from '../actions/settingsAction';
+import { ThunkDispatch } from 'redux-thunk';
 
 const mapStateToProps = (state: AppState) => ({
   isLoggedIn: !!state.authStatus.userId,
+  loading: !state.authStatus.checked,
+  alreadyLoaded: state.authStatus.checked,
 });
-const mapDispatchToProps = (dispatch: Dispatch) => ({});
+const mapDispatchToProps = (
+  dispatch: ThunkDispatch<AppState, undefined, AnyAction>,
+) => ({
+  load: () => {
+    dispatch(fetchSettings());
+  },
+});
 
 export default connect(
   mapStateToProps,
